#50 refactor: 位置表示コンポーネントを現在位置・予定位置に分離する
- ActorMarker (position + path + trail 統合) を廃止し、
  CurrentPositionMarker (positionStore のみ購読)、
  PlannedPathMarker (plannedPathStore のみ購読) に分離
- set target では PlannedPathMarker のみ、行動決定では
  CurrentPositionMarker のみ再レンダリングされるようになる
- 軌跡 (trail) 表示は今回のスコープ外として廃止。専用だった
  getEventLogByActorId も未使用のため削除
